+++ spec/lib/karafka/connection/messages_buffer_spec.rb
@@ -97,4 +97,79 @@ RSpec.describe_current do
       expect { |block| buffer.each(&block) }.not_to yield_control
     end
   end
+
+  describe '#delete' do
+    let(:message1) { build(:messages_message) }
+    let(:message2) { build(:messages_message) }
+
+    before { buffer << message1 }
+
+    context 'when given topic does not exist' do
+      it 'expect not to remove anything' do
+        expect { buffer.delete('na', 0) }.not_to change(buffer, :size)
+      end
+    end
+
+    context 'when given partition does not exist' do
+      it 'expect not to remove anything' do
+        expect { buffer.delete(message1.topic, 100) }.not_to change(buffer, :size)
+      end
+    end
+
+    context 'when given partition exists and is the only one' do
+      it 'expect change the buffer size to reflect the change' do
+        expect { buffer.delete(message1.topic, 0) }.to change(buffer, :size).by(-1)
+      end
+
+      it 'expect to completely remove the topic when no other partitions present' do
+        buffer.delete(message1.topic, 0)
+
+        expect { |block| buffer.each(&block) }.not_to yield_control
+      end
+    end
+
+    context 'when given partition exists and is not the only one' do
+      let(:message2) { OpenStruct.new(topic: message1.topic, partition: 1) }
+
+      before { buffer << message2 }
+
+      it 'expect change the buffer size to reflect the change' do
+        expect { buffer.delete(message1.topic, 0) }.to change(buffer, :size).by(-1)
+      end
+
+      it 'expect to recalculate the buffer size' do
+        expect(buffer.size).to eq(2)
+        buffer.delete(message1.topic, 0)
+        expect(buffer.size).to eq(1)
+      end
+
+      it 'expect not to mix the topics' do
+        buffer.delete(message1.topic, 0)
+
+        buffer.each do |topic, partition, messages|
+          expect(topic).to eq(message2.topic)
+        end
+      end
+
+      it 'expect not to mix partitions' do
+        buffer.delete(message1.topic, 0)
+
+        buffer.each do |topic, partition, messages|
+          expect(partition).to eq(message2.partition)
+        end
+      end
+
+      it 'expect not to mix messages in the buffer' do
+        buffer.delete(message1.topic, 0)
+
+        buffer.each do |topic, partition, messages|
+          expect(messages).to eq([message2])
+        end
+      end
+    end
+  end
+
+  describe 'uniq!' do
+    pending
+  end
 end