+++ spec/lib/karafka/connection/rebalance_manager_spec.rb
@@ -3,42 +3,20 @@
 RSpec.describe_current do
   subject(:manager) { described_class.new }
 
-  let(:partition) { Rdkafka::Consumer::Partition.new(rand(0..100), 'topic_name') }
-  let(:partitions) { { 'topic_name' => [partition] } }
-
-  describe '#assigned_partitions and #on_partitions_assigned' do
-    context 'when there are no assigned partitions' do
-      it { expect(manager.assigned_partitions).to eq({}) }
-    end
-
-    context 'when some partitions are assigned' do
-      before { manager.on_partitions_assigned(nil, partitions) }
-
-      it 'expect to return them' do
-        expect(manager.assigned_partitions).to eq({ 'topic_name' => [partition.partition] })
-      end
-    end
-
-    context 'when we clear the manager' do
-      before do
-        manager.on_partitions_assigned(nil, partitions)
-        manager.clear
-      end
-
-      it { expect(manager.assigned_partitions).to eq({}) }
-    end
-  end
+  let(:partition1) { Rdkafka::Consumer::Partition.new(1, 'topic_name') }
+  let(:partition2) { Rdkafka::Consumer::Partition.new(4, 'topic_name') }
+  let(:partitions) { { 'topic_name' => [partition1] } }
 
   describe '#revoked_partitions and #on_partitions_revoked' do
     context 'when there are no revoked partitions' do
       it { expect(manager.revoked_partitions).to eq({}) }
     end
 
-    context 'when some partitions are revoked' do
+    context 'when some partitions were revoked and not assigned' do
       before { manager.on_partitions_revoked(nil, partitions) }
 
       it 'expect to return them' do
-        expect(manager.revoked_partitions).to eq({ 'topic_name' => [partition.partition] })
+        expect(manager.revoked_partitions).to eq({ 'topic_name' => [partition1.partition] })
       end
     end
 
@@ -49,6 +27,17 @@ RSpec.describe_current do
       end
 
       it { expect(manager.revoked_partitions).to eq({}) }
+    end
+
+    context 'when some of the revoked partitions were assigned back' do
+      before do
+        manager.on_partitions_assigned(nil, { 'topic_name' => [partition1] })
+        manager.on_partitions_revoked(nil, { 'topic_name' => [partition1, partition2] })
+      end
+
+      it 'expect not to include them in the revoked partitions back' do
+        expect(manager.revoked_partitions).to eq({ 'topic_name' => [partition2.partition] })
+      end
     end
   end
 end